Use uint8_t for struct ip bit fields
Use uint8_t instead of unsigned for the ip_hl and ip_v fields in
struct ip so that they'll occupy one byte instead of four on Windows.

(MSVC allocates space sufficient for the bit field type rather than the
field width.  This corresponds to GCC's -mms-bitfields option, which is
enabled by default on Windows.  On other platforms, GCC allocate space
sufficient for the field width regardless of type.)